recommendation_service/requirements.txt
fastapi
uvicorn
pymongo
scikit-learn
pandas
python-dotenv
